Moved AddEditActivity.java to a new folder
