+++ app/src/main/res/layout/activity_add_edit_vibe.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddEditVibeActivity" >
+    tools:context=".AddEdit.AddEditVibeActivity" >
 
     <LinearLayout
         android:layout_width="match_parent"
