+++ docs/javadocs/com/example/vybe/AddEditVibeActivity.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>com.example.vybe.AddEditVibeActivity</li>
+<li>com.example.vybe.AddEdit.AddEditVibeActivity</li>
 </ul>
 </li>
 </ul>
